+++ src/wdio/services/index.js
@@ -1,0 +1,9 @@
+import AxeService from './AxeService';
+import TerraService from './TerraService';
+import SeleniumDockerService from './SeleniumDockerService';
+import WebpackDevServerService from './WebpackDevServerService';
+
+module.exports.Axe = new AxeService();
+module.exports.Terra = new TerraService();
+module.exports.SeleniumDocker = new SeleniumDockerService();
+module.exports.WebpackDevService = new WebpackDevServerService();